# src/main/resources/api/openapi.yaml
openapi: 3.0.2
info:
  description: Logistics API provides services to find and manage entities for logistics
  title: Logistics API
  version: "1.0"
  contact:
    email: vladislavprokopenko1@gmail.com
servers:
  - url: http://localhost:8080
...
paths:
  /trucks:
    get:
      tags:
        - Truck
      summary: Find all Trucks
      operationId: trucksFindAll
      responses:
        200:
          description: Trucks were found successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Truck'
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        404:
          description: Trucks were not found
        500:
          description: Internal Server Error

    post:
      tags:
        - Truck
      summary: Create a new Truck
      operationId: truckCreate
      requestBody:
        description: Truck object that needs to be added to truck list
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTruck'
      responses:
        201:
          description: Truck was created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Truck'
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        500:
          description: Internal Server Error
  /trucks/byLogistician/{number}:
    patch:
      tags:
        - Truck
      summary: Update Truck
      operationId: truckUpdateByLogistician
      parameters:
        - name: number
          in: path
          description: "Truck number in proper format"
          required: true
          schema:
            type: string
            example: ABC1234
      requestBody:
        description: Truck object that needs to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTruckByLogistician'
      responses:
        200:
          description: Truck was updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Truck'
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        500:
          description: Internal Server Error
  /trucks/byDriver/{number}:
    patch:
      tags:
        - Truck
      summary: Update Truck
      operationId: truckUpdateByDriver
      parameters:
        - name: number
          in: path
          description: "Truck number in proper format"
          required: true
          schema:
            type: string
            example: ABC1234
      requestBody:
        description: Truck object that needs to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTruckByDriver'
      responses:
        200:
          description: Truck was updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Truck'
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        500:
          description: Internal Server Error
  /trucks/{number}:
    delete:
      tags:
        - Truck
      summary: Delete Truck
      operationId: truckDelete
      parameters:
        - name: number
          in: path
          description: "Truck number in proper format"
          required: true
          schema:
            type: string
            example: ABC1234
      responses:
        204:
          description: Truck was deleted successfully
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        404:
          description: This truck number was not found
        500:
          description: Internal Server Error

    get:
      tags:
        - Truck
      summary: Find Truck by number
      operationId: truckFindByNumber
      parameters:
        - name: number
          in: path
          description: "Truck number in proper format"
          required: true
          schema:
            type: string
            example: ABC1234
      responses:
        200:
          description: Truck was found successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Truck'
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        404:
          description: This truck number was not found
        500:
          description: Internal Server Error
  /trucks/technicalCondition/{technicalConditionStatus}:
    get:
      tags:
        - Truck
      summary: Find Trucks by technical condition
      operationId: trucksFindByTechnicalCondition
      parameters:
        - name: technicalConditionStatus
          in: path
          description: "Technical condition can be: OK / NOK"
          required: true
          schema:
            type: string
            example: OK
      responses:
        200:
          description: Trucks in this technical condition were found successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Truck'
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        404:
          description: Trucks in this technical condition were not found
        500:
          description: Internal Server Error
  /trucks/busy/{busyStatus}:
    get:
      tags:
        - Truck
      summary: Find Trucks by busy status
      operationId: trucksFindByBusyStatus
      parameters:
        - name: busyStatus
          in: path
          description: "Busy status can be: YES / NO"
          allowEmptyValue: false
          required: true
          schema:
            type: string
            enum: ["YES","NO"]
            example: "NO"
      responses:
        200:
          description: Trucks with defined busy status were found successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Truck'
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        404:
          description: Trucks with defined busy status were not found
        500:
          description: Internal Server Error
  /trucks/currentCityAndState:
    get:
      tags:
        - Truck
      summary: Find Trucks by current city and state
      operationId: trucksFindByCurrentCityAndState
      parameters:
        - in: query
          name: city
          description: Current city
          allowEmptyValue: false
          required: true
          schema:
            type: string
            example: Chicago
          allowReserved: false
        - in: query
          name: state
          description: Current state
          allowEmptyValue: false
          required: true
          schema:
            type: string
            example: Illinois
          allowReserved: false
      responses:
        200:
          description: Trucks were found successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Truck'
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        404:
          description: Trucks in this city and state were not found
        500:
          description: Internal Server Error
  /trucks/forOrder:
    get:
      tags:
        - Truck
      summary: "Find proper Trucks for Order: not busy, in current city and state, having required capacity"
      operationId: trucksFindForOrder
      parameters:
        - in: query
          name: orderId
          description: "Order id"
          required: true
          schema:
            type: integer
            format: int32
            example: 12001
        - in: query
          name: city
          description: Current city
          allowEmptyValue: false
          required: true
          schema:
            type: string
            example: Chicago
          allowReserved: false
        - in: query
          name: state
          description: Current state
          allowEmptyValue: false
          required: true
          schema:
            type: string
            example: Illinois
          allowReserved: false
        - in: query
          name: capacity
          description: Minimum required Truck capacity
          allowEmptyValue: false
          required: true
          schema:
            type: number
            format: double
            example: 15.4
          allowReserved: false
      responses:
        200:
          description: Trucks were found successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Truck'
        400:
          description: Bad request
        401:
          description: "Unauthorized: Invalid credentials"
        403:
          description: "Forbidden: User does not have permissions"
        404:
          description: Proper Trucks for this Order were not found
        500:
          description: Internal Server Error
...
components:
  schemas:
    CreateTruck:
      description: Object representing Truck creating
      type: object
      properties:
        number:
          type: string
          example: ABC1234
        model:
          type: string
          example: "Peterbilt 358"
        capacity:
          description: Truck capacity, tons
          type: number
          format: double
          example: 15.4
        totalOperatingTime:
          description: Total operating time, years
          type: integer
          format: int32
          minimum: 0
          maximum: 50
          example: 10
        currentCity:
          type: string
          example: Chicago
        currentState:
          type: string
          example: Illinois
    UpdateTruckByLogistician:
      description: Object representing Truck updating by Logistician
      type: object
      properties:
        model:
          type: string
          example: "Peterbilt 358"
        capacity:
          description: Truck capacity, tons
          type: number
          format: double
          example: 15.4
        totalOperatingTime:
          description: Total operating time, years
          type: integer
          format: int32
          minimum: 0
          maximum: 50
          example: 10
        technicalCondition:
          type: string
          enum: [OK,NOK]
          default: OK
        currentCity:
          type: string
          example: Chicago
        currentState:
          type: string
          example: Illinois
    UpdateTruckByDriver:
      description: Object representing Truck updating by Driver
      type: object
      properties:
        technicalCondition:
          type: string
          enum: [OK,NOK]
          default: OK
    Truck:
      description: Object representing Truck
      type: object
      properties:
        number:
          type: string
          example: ABC1234
        model:
          type: string
          example: "Peterbilt 358"
        capacity:
          description: Truck capacity, tons
          type: number
          format: double
          example: 15.4
        totalOperatingTime:
          description: Total operating time, years
          type: integer
          format: int32
          minimum: 0
          maximum: 50
          example: 10
        technicalCondition:
          type: string
          enum: [OK,NOK]
          default: OK
        busy:
          description: Busy status
          type: string
          enum: ["YES","NO"]
          default: "NO"
        currentCity:
          type: string
          example: Chicago
        currentState:
          type: string
          example: Illinois
...